requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pip==23.0.1
PyJWT==2.8.0
setuptools==65.5.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
crispy-bootstrap5
django-crispy-forms
markdown
django-extensions
django-cors-headers
six==1.16.0
webencodings==0.5.1
bleach==6.1.0
django-crum==0.7.9
